remove store. Remove exitButtonComponent
